crwpr/requirements.txt
llama-index-core
llama-index-readers-web
playwright
beautifulsoup4